backend/requirements.txt

Django==3.2.3
django-cleanup==8.1.0
django-colorfield==0.11.0
django-cors-headers==3.13.0
django-filter==23.5
djangorestframework==3.12.4
djoser==2.1.0
drf-extra-fields ==3.7.0
gunicorn==20.1.0
Pillow==10.2.0
psycopg2-binary==2.9.3
python-dotenv==1.0.1
PyYAML==6.0
reportlab==4.1.0
requests==2.31.0
webcolors==1.11.1
